Added ref and ref tests

+++ lang/reference.c
@@ -2,7 +2,9 @@
 #include <stdlib.h>
 #include "reference.h"
 
-ref* new_ref(const u32 t, const void *e){
+#include <stdio.h>
+
+ref* new_ref(const u8 t, const void *e){
   const ref mref = {t,e};           // mref - mutable reference
   ref *rf = malloc(sizeof(ref));    // allocate a pointer for the return
   memcpy(rf, &mref, sizeof(ref));   // memcpy in order to init a malloc'd reference
